updated docs.

+++ Arrays & Strings/Arrays/helpers.h
@@ -2,10 +2,10 @@
 #define ARRAY_HELPERS
 #include <stdio.h>
 
-const int MATRIX_M = 6;
-const int MATRIX_N = 4;
+const size_t MATRIX_M = 6;
+const size_t MATRIX_N = 4;
 
-void matrixMxNFill(int matrix[MATRIX_M][MATRIX_N], int m_count, int n_count, int value) {
+void matrixMxNFill(int matrix[MATRIX_M][MATRIX_N], size_t m_count, size_t n_count, int value) {
     for (int i = 0; i < m_count; ++i) {
         for (int j = 0; j < n_count; ++j) {
             matrix[i][j] = value;
@@ -13,7 +13,7 @@ void matrixMxNFill(int matrix[MATRIX_M][MATRIX_N], int m_count, int n_count, int
     }
 }
 
-void printIntMxNMatrix(int matrix[MATRIX_M][MATRIX_N], int m_count, int n_count) {
+void printIntMxNMatrix(int matrix[MATRIX_M][MATRIX_N], size_t m_count, size_t n_count) {
     for (int i = 0; i < m_count; ++i) {
         for (int j = 0; j < n_count; ++j) {
             printf("%d ", matrix[i][j]);
@@ -22,7 +22,7 @@ void printIntMxNMatrix(int matrix[MATRIX_M][MATRIX_N], int m_count, int n_count)
     }
 }
 
-void printIntNxNMatrix(int matrix[MATRIX_N][MATRIX_N], int n_count) {
+void printIntNxNMatrix(int matrix[MATRIX_N][MATRIX_N], size_t n_count) {
     for (int i = 0; i < n_count; ++i) {
         for (int j = 0; j < n_count; ++j) {
             printf("%d ", matrix[i][j]);
