+++ Arrays & Strings/Strings/checkPermutation.c
@@ -2,7 +2,9 @@
 #include "helpers.h"
 
 /**
- * O(n)
+ * Time: O(n)
+ * Space: O(n)
+ * 
  * Very similar to isUnique 
  * 
  * Build a poor mans Hash Table for all ASCII values at the character code
@@ -10,7 +12,7 @@
  * Decrement the count at the character codes index in the Hash Table for string two
  * If at the end of this process the Hash Table has 0 for very index the strings are permutations of each other
  */
-int checkPermutation(char s1[], char s2[], int size) {
+int checkPermutation(char s1[], char s2[], size_t size) {
     int arr[ASCII_CHAR_EXTENDED];
     fill(arr, ASCII_CHAR_EXTENDED, 0);
 
@@ -23,16 +25,12 @@ int checkPermutation(char s1[], char s2[], int size) {
 }
 
 /**
- * O(n)
  * Give two strings, write a method to decide if one is a permutation of the other.
- * 
- * I'm assuming both strings are the same length becuase the prompt did not ask
- * if the permutation is a substring of the other
  */
 int main() {
     char s1[] = "abdb";
     char s2[] = "bbda";
-    int size = sizeof(s1) / sizeof(char);
+    size_t size = sizeof(s1) / sizeof(char);
     printf("%d\n", checkPermutation(s1, s2, size));
     return 0;
 }
