+++ Arrays & Strings/Strings/urlify.c
@@ -6,7 +6,7 @@
  * Space: O(c)
  * n = Length of string
  */
-void urlify(char s[], int s_count) {
+void urlify(char s[], size_t s_count) {
     int spaces = char_count(s, s_count, ' ');
     int total = s_count + (spaces * 2);
 
@@ -49,7 +49,7 @@ void urlify(char s[], int s_count) {
  */
 int main() {
     char s[] = "Mr John Smith    ";
-    int size = 13;
+    size_t size = 13;
     urlify(s, size);
     printf("%s\n", s);
     return 0;
